# march-madness-stats/backend/src/main/resources/schemas/swagger.json
{
    "openapi": "3.0.0",
    "components": {
...
        "schemas": {
...
            "SeasonType": {
                "enum": [
                    "postseason",
                    "preseason",
                    "regular"
                ],
                "type": "string"
            },
...
            "GameStatus": {
                "enum": [
                    "scheduled",
                    "in_progress",
                    "final",
                    "postponed",
                    "cancelled"
                ],
                "type": "string"
            },
            "GameInfo": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "sourceId": {
                        "type": "string"
                    },
                    "seasonLabel": {
                        "type": "string"
                    },
                    "season": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "seasonType": {
                        "$ref": "#/components/schemas/SeasonType"
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "startTimeTbd": {
                        "type": "boolean"
                    },
                    "neutralSite": {
                        "type": "boolean"
                    },
                    "conferenceGame": {
                        "type": "boolean"
                    },
                    "gameType": {
                        "type": "string",
                        "nullable": true
                    },
                    "tournament": {
                        "type": "string",
                        "nullable": true
                    },
                    "gameNotes": {
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "$ref": "#/components/schemas/GameStatus"
                    },
                    "attendance": {
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "homeTeamId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "homeTeam": {
                        "type": "string"
                    },
                    "homeConferenceId": {
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "homeConference": {
                        "type": "string",
                        "nullable": true
                    },
                    "homeSeed": {
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "homePoints": {
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "homePeriodPoints": {
                        "items": {
                            "type": "integer",
                            "format": "int32"
                        },
                        "type": "array",
                        "nullable": true
                    },
                    "homeWinner": {
                        "type": "boolean",
                        "nullable": true
                    },
                    "awayTeamId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "awayTeam": {
                        "type": "string"
                    },
                    "awayConferenceId": {
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "awayConference": {
                        "type": "string",
                        "nullable": true
                    },
                    "awaySeed": {
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "awayPoints": {
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "awayPeriodPoints": {
                        "items": {
                            "type": "integer",
                            "format": "int32"
                        },
                        "type": "array",
                        "nullable": true
                    },
                    "awayWinner": {
                        "type": "boolean",
                        "nullable": true
                    },
                    "excitement": {
                        "type": "number",
                        "format": "double",
                        "nullable": true
                    },
                    "venueId": {
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "venue": {
                        "type": "string",
                        "nullable": true
                    },
                    "city": {
                        "type": "string",
                        "nullable": true
                    },
                    "state": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "required": [
                    "id",
                    "sourceId",
                    "seasonLabel",
                    "season",
                    "seasonType",
                    "startDate",
                    "startTimeTbd",
                    "neutralSite",
                    "conferenceGame",
                    "gameType",
                    "tournament",
                    "gameNotes",
                    "status",
                    "attendance",
                    "homeTeamId",
                    "homeTeam",
                    "homeConferenceId",
                    "homeConference",
                    "homeSeed",
                    "homePoints",
                    "homePeriodPoints",
                    "homeWinner",
                    "awayTeamId",
                    "awayTeam",
                    "awayConferenceId",
                    "awayConference",
                    "awaySeed",
                    "awayPoints",
                    "awayPeriodPoints",
                    "awayWinner",
                    "excitement",
                    "venueId",
                    "venue",
                    "city",
                    "state"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "GameMediaInfo": {
                "properties": {
                    "gameId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "season": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "seasonLabel": {
                        "type": "string"
                    },
                    "seasonType": {
                        "$ref": "#/components/schemas/SeasonType"
                    },
                    "tournament": {
                        "type": "string",
                        "nullable": true
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "startTimeTbd": {
                        "type": "boolean"
                    },
                    "homeTeamId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "homeTeam": {
                        "type": "string"
                    },
                    "homeConference": {
                        "type": "string",
                        "nullable": true
                    },
                    "awayTeamId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "awayTeam": {
                        "type": "string"
                    },
                    "awayConference": {
                        "type": "string",
                        "nullable": true
                    },
                    "neutralSite": {
                        "type": "boolean"
                    },
                    "conferenceGame": {
                        "type": "boolean"
                    },
                    "gameType": {
                        "type": "string",
                        "nullable": true
                    },
                    "notes": {
                        "type": "string",
                        "nullable": true
                    },
                    "broadcasts": {
                        "items": {
                            "properties": {
                                "broadcastName": {
                                    "type": "string"
                                },
                                "broadcastType": {
                                    "type": "string"
                                }
                            },
                            "required": [
                                "broadcastName",
                                "broadcastType"
                            ],
                            "type": "object"
                        },
                        "type": "array"
                    }
                },
                "required": [
                    "gameId",
                    "season",
                    "seasonLabel",
                    "seasonType",
                    "tournament",
                    "startDate",
                    "startTimeTbd",
                    "homeTeamId",
                    "homeTeam",
                    "homeConference",
                    "awayTeamId",
                    "awayTeam",
                    "awayConference",
                    "neutralSite",
                    "conferenceGame",
                    "gameType",
                    "notes",
                    "broadcasts"
                ],
                "type": "object",
                "additionalProperties": false
            },
...
            "GameBoxScoreTeam": {
                "properties": {
                    "gameId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "season": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "seasonLabel": {
                        "type": "string"
                    },
                    "seasonType": {
                        "$ref": "#/components/schemas/SeasonType"
                    },
                    "tournament": {
                        "type": "string",
                        "nullable": true
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "startTimeTbd": {
                        "type": "boolean"
                    },
                    "teamId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "team": {
                        "type": "string"
                    },
                    "conference": {
                        "type": "string",
                        "nullable": true
                    },
                    "teamSeed": {
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "opponentId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "opponent": {
                        "type": "string"
                    },
                    "opponentConference": {
                        "type": "string",
                        "nullable": true
                    },
                    "opponentSeed": {
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "neutralSite": {
                        "type": "boolean"
                    },
                    "isHome": {
                        "type": "boolean"
                    },
                    "conferenceGame": {
                        "type": "boolean"
                    },
                    "gameType": {
                        "type": "string",
                        "nullable": true
                    },
                    "notes": {
                        "type": "string",
                        "nullable": true
                    },
                    "gameMinutes": {
                        "type": "number",
                        "format": "double",
                        "nullable": true
                    },
                    "pace": {
                        "type": "number",
                        "format": "double",
                        "nullable": true
                    },
                    "teamStats": {
                        "$ref": "#/components/schemas/GameBoxScoreTeamStats"
                    },
                    "opponentStats": {
                        "$ref": "#/components/schemas/GameBoxScoreTeamStats"
                    }
                },
                "required": [
                    "gameId",
                    "season",
                    "seasonLabel",
                    "seasonType",
                    "tournament",
                    "startDate",
                    "startTimeTbd",
                    "teamId",
                    "team",
                    "conference",
                    "teamSeed",
                    "opponentId",
                    "opponent",
                    "opponentConference",
                    "opponentSeed",
                    "neutralSite",
                    "isHome",
                    "conferenceGame",
                    "gameType",
                    "notes",
                    "gameMinutes",
                    "pace",
                    "teamStats",
                    "opponentStats"
                ],
                "type": "object",
                "additionalProperties": false
            },
...
        },
        "securitySchemes": {
            "apiKey": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "info": {
        "title": "College Basketball Data API",
        "version": "1.20.1",
        "description": "This API is in limited Beta for Patreon subscribers. It may have bugs and is subject to changes. API keys can be acquired from the CollegeFootballData.com website.",
        "license": {
            "name": "MIT"
        },
        "contact": {
            "email": "admin@collegefootballdata.com",
            "name": "the developer"
        }
    },
    "paths": {
...
        "/games": {
            "get": {
                "operationId": "GetGames",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/GameInfo"
                                    },
                                    "type": "array"
                                }
                            }
                        }
                    }
                },
                "description": "Returns information on the first 3000 games that match the provided filters, ordered by start date.",
                "tags": [
                    "games"
                ],
                "security": [
                    {
                        "apiKey": []
                    }
                ],
                "parameters": [
                    {
                        "description": "Optional start timestamp in ISO 8601 format",
                        "in": "query",
                        "name": "startDateRange",
                        "required": false,
                        "schema": {
                            "format": "date-time",
                            "type": "string"
                        }
                    },
                    {
                        "description": "Optional end timestamp in ISO 8601 format",
                        "in": "query",
                        "name": "endDateRange",
                        "required": false,
                        "schema": {
                            "format": "date-time",
                            "type": "string"
                        }
                    },
                    {
                        "description": "Optional team name filter",
                        "in": "query",
                        "name": "team",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Optional conference abbreviation filter",
                        "in": "query",
                        "name": "conference",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Optional season filter",
                        "in": "query",
                        "name": "season",
                        "required": false,
                        "schema": {
                            "format": "int32",
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Optional season type filter",
                        "in": "query",
                        "name": "seasonType",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/SeasonType"
                        }
                    },
                    {
                        "description": "Optional game status filter",
                        "in": "query",
                        "name": "status",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/GameStatus"
                        }
                    },
                    {
                        "description": "Optional tournament filter (e.g. NCAA, NIT, etc)",
                        "in": "query",
                        "name": "tournament",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/games/media": {
            "get": {
                "operationId": "GetBroadcasts",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/GameMediaInfo"
                                    },
                                    "type": "array"
                                }
                            }
                        }
                    }
                },
                "description": "Returns broadcast information on the first 3000 games that match the provided filters, ordered by start date.",
                "tags": [
                    "games"
                ],
                "security": [
                    {
                        "apiKey": []
                    }
                ],
                "parameters": [
                    {
                        "description": "Optional start timestamp in ISO 8601 format",
                        "in": "query",
                        "name": "startDateRange",
                        "required": false,
                        "schema": {
                            "format": "date-time",
                            "type": "string"
                        }
                    },
                    {
                        "description": "Optional end timestamp in ISO 8601 format",
                        "in": "query",
                        "name": "endDateRange",
                        "required": false,
                        "schema": {
                            "format": "date-time",
                            "type": "string"
                        }
                    },
                    {
                        "description": "Optional team name filter",
                        "in": "query",
                        "name": "team",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Optional conference abbreviation filter",
                        "in": "query",
                        "name": "conference",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Optional season filter",
                        "in": "query",
                        "name": "season",
                        "required": false,
                        "schema": {
                            "format": "int32",
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Optional season type filter",
                        "in": "query",
                        "name": "seasonType",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/SeasonType"
                        }
                    },
                    {
                        "description": "Optional tournament filter (e.g. NCAA, NIT, etc)",
                        "in": "query",
                        "name": "tournament",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/games/teams": {
            "get": {
                "operationId": "GetGameTeams",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/GameBoxScoreTeam"
                                    },
                                    "type": "array"
                                }
                            }
                        }
                    }
                },
                "description": "Returns team box score statistics and metrics on the first 3000 games that match the provided filters, ordered by start date.",
                "tags": [
                    "games"
                ],
                "security": [
                    {
                        "apiKey": []
                    }
                ],
                "parameters": [
                    {
                        "description": "Optional start timestamp in ISO 8601 format",
                        "in": "query",
                        "name": "startDateRange",
                        "required": false,
                        "schema": {
                            "format": "date-time",
                            "type": "string"
                        }
                    },
                    {
                        "description": "Optional end timestamp in ISO 8601 format",
                        "in": "query",
                        "name": "endDateRange",
                        "required": false,
                        "schema": {
                            "format": "date-time",
                            "type": "string"
                        }
                    },
                    {
                        "description": "Optional team name filter",
                        "in": "query",
                        "name": "team",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Optional conference abbreviation filter",
                        "in": "query",
                        "name": "conference",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Optional season filter",
                        "in": "query",
                        "name": "season",
                        "required": false,
                        "schema": {
                            "format": "int32",
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Optional season type filter",
                        "in": "query",
                        "name": "seasonType",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/SeasonType"
                        }
                    },
                    {
                        "description": "Optional tournament filter (e.g. NCAA, NIT, etc)",
                        "in": "query",
                        "name": "tournament",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/games/players": {
            "get": {
                "operationId": "GetGamePlayers",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/GameBoxScorePlayers"
                                    },
                                    "type": "array"
                                }
                            }
                        }
                    }
                },
                "description": "Returns player box score statistics and metrics on the first 1000 games that match the provided filters, ordered by start date.",
                "tags": [
                    "games"
                ],
                "security": [
                    {
                        "apiKey": []
                    }
                ],
                "parameters": [
                    {
                        "description": "Optional start timestamp in ISO 8601 format",
                        "in": "query",
                        "name": "startDateRange",
                        "required": false,
                        "schema": {
                            "format": "date-time",
                            "type": "string"
                        }
                    },
                    {
                        "description": "Optional end timestamp in ISO 8601 format",
                        "in": "query",
                        "name": "endDateRange",
                        "required": false,
                        "schema": {
                            "format": "date-time",
                            "type": "string"
                        }
                    },
                    {
                        "description": "Optional team name filter",
                        "in": "query",
                        "name": "team",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Optional conference abbreviation filter",
                        "in": "query",
                        "name": "conference",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Optional season filter",
                        "in": "query",
                        "name": "season",
                        "required": false,
                        "schema": {
                            "format": "int32",
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Optional season type filter",
                        "in": "query",
                        "name": "seasonType",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/SeasonType"
                        }
                    },
                    {
                        "description": "Optional tournament filter (e.g. NCAA, NIT, etc)",
                        "in": "query",
                        "name": "tournament",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
...
    },
    "servers": [
        {
            "url": "https://api.collegebasketballdata.com/"
        }
    ],
...
}